requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.9.1
Bootstrap-Flask==2.2.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
flake8==6.0.0
Flask==2.2.2
fonttools==4.39.0
frozenlist==1.3.3
gpxpy==1.5.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
magic-filter==1.0.9
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
multidict==6.0.4
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
polyline==2.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.2
